pycalc/requirements.txt
altgraph==0.16.1
future==0.16.0
gevent==1.3.7
greenlet==0.4.15
macholib==1.11
msgpack-python==0.5.6
pefile==2018.8.8
PyInstaller==3.4
pyzmq==17.1.2
zerorpc==0.6.1
